refactor(cli): extract session navigation into CLISessionNavigationMixin (god-file Phase 3)
CLICommandsMixin is itself a product of the god-file decomposition campaign,
lifted out of cli.py. It has since grown to 3,218 lines and 54 methods, which
makes it a decomposition target of its own.

This takes the one cluster in it that is about moving between sessions rather
than acting inside the current one: handing the session off to a gateway
platform, resuming an earlier one, listing what can be resumed, and forking a
branch.

Moved verbatim (531 lines, 4 methods, contiguous at 707-1240):

  _handle_handoff_command     _handle_sessions_command
  _handle_resume_command      _handle_branch_command

Behavior-neutral. Every body is byte-identical to what it replaced; the only
edits are the class declaration, one import, and the new module header.

The new mixin becomes a base of CLICommandsMixin rather than of HermesCLI, so
cli.py is untouched and the change stays inside one file plus the new module.
HermesCLI picks the methods up through the existing MRO: HermesCLI ->
CLIAgentSetupMixin -> CLICommandsMixin -> CLISessionNavigationMixin ->
CLIBillingMixin. Verified at runtime that every method name defined on the old
class still resolves off HermesCLI.

The lifted methods reach cli module state the way they already did, through
lazy imports inside the bodies (from cli import _cprint, from cli import
_sync_process_session_id). Those resolve from cli's namespace at call time,
which is what keeps patch("cli._cprint") working from the new module exactly
as before. Only four module-level names had to follow: os, uuid, datetime and
extract_api_content_sidecar, taken from their original sources.

cli_commands_mixin.py 3,260 -> 2,726 lines; the class 3,218 -> 2,683 lines and
54 -> 50 methods.
